requirements.txt
# requirements.txt

torch
transformers
streamlit
scikit-learn
numpy
Pillow
datasets